2477 userguide update

+++ documents/userguide/scripts/4-hosted-download.sh
@@ -1,21 +1,19 @@
-#!/bin/sh -ex
-
-printenv DOMAIN > /dev/null
+#!/bin/bash
+set -e
 
 # tag::public[]
 id=$1
 
 curl -S -s -X GET \
+    -u "$PZUSER":"$PZPASS" \
     -w "%{http_code}" \
     -o response.txt \
     "https://pz-gateway.$DOMAIN/file/$id?fileName=terrametrics.tif" > status.txt
 
 # verify all worked successfully
-grep -q 200 status.txt
-grep -q jobId response.txt
+grep -q 200 status.txt || { cat response.txt; exit 1; }
+# end::public[]
 
-# print out the JobId
-grep -o '"jobId":".*"' response.txt | cut -d \" -f 4
-# end::public[] 
+echo pass.
 
 rm -f response.txt status.txt
